+++ src/symlink.cpp
@@ -84,7 +84,7 @@ int
 _symlink(Policy::Func::Search  searchFunc,
          Policy::Func::Create  createFunc,
          const vector<string> &srcmounts,
-         const size_t          minfreespace,
+         const uint64_t        minfreespace,
          const char           *oldpath,
          const char           *newpath)
 {
